pyproject.toml
[tool.poetry]
name = "automated-attendance-website"
version = "0.1.0"
description = ""
authors = ["Hammad Shami <H2Shami@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.0"
uvicorn = "^0.27.1"
psycopg2 = "^2.9.9"
boto3 = "^1.34.59"
python-dotenv = "^1.0.1"
python-multipart = "^0.0.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
